+++ assets/js/clippi-config.js
@@ -1,6 +1,6 @@
 (function(){
   window.CLIPPIO_CONFIG = {
-    version: '6.6.0',
+    version: '6.6.1',
     avatarUrl: '/assets/images/clippi-avatar.png',
     webFinderUrl: '/web-finder/',
     contactUrl: '/kontakt/',